Rename code block utils file
It's no longer only about finding a code block from the channel.
